Add jlambda macro
As we are now getting the field/method attributes (modifiers, types,
etc) in runtime, I need to modify jlambda-field (ready) and
jlambda-method (pending) to work with this attributes at runtime instead
of expansion time.
To maintain the user API, jlambda-field has been
splitted in jlambda-field* which is invoked by jlambda, and
jlambda-field which kept the previous semantic.